refactor paimon arrow converter layout
