+++ be/test/core/data_type_serde/data_type_variant_v2_serde_output_test.cpp
@@ -47,8 +47,8 @@
 #include "core/data_type_serde/data_type_variant_v2_serde.h"
 #include "core/string_buffer.hpp"
 #include "core/value/variant/variant_parquet_encoding.h"
-#include "exec/sink/writer/paimon/paimon_arrow_write_converter.h"
 #include "exprs/function/parse/variant_string_parse.h"
+#include "format/table/paimon/paimon_arrow_write_converter.h"
 #include "util/mysql_row_buffer.h"
 
 namespace doris {
@@ -223,7 +223,7 @@ void expect_paimon_variant_bytes(const IColumn& column, const ColumnVariantV2& e
     DataTypePtr type = std::make_shared<DataTypeVariantV2>();
     const auto serde = type->get_serde();
     const auto field = arrow::field("payload", binary_variant_arrow_type(), true);
-    const Status status = paimon_arrow_write_converter().write_column(
+    const Status status = paimon::paimon_arrow_write_converter().write_column(
             type, *serde, column, null_map, field, builder.get(), 0, column.size(),
             cctz::utc_time_zone());
     ASSERT_TRUE(status.ok()) << status;
@@ -477,7 +477,7 @@ TEST(PaimonArrowWriteConverterTest, NestedArrayUsesPaimonSerdeRecursively) {
     ASSERT_TRUE(arrow::MakeBuilder(arrow::default_memory_pool(), arrow_type, &builder).ok());
 
     const auto serde = array_type->get_serde();
-    Status status = paimon_arrow_write_converter().write_column(
+    Status status = paimon::paimon_arrow_write_converter().write_column(
             array_type, *serde, *array, nullptr, field, builder.get(), 0, array->size(),
             cctz::utc_time_zone());
     ASSERT_TRUE(status.ok()) << status;
@@ -511,7 +511,7 @@ TEST(PaimonArrowWriteConverterTest, BinaryStructRejectsUnsupportedPaimonPrimitiv
     DataTypePtr type = std::make_shared<DataTypeVariantV2>();
     const auto serde = type->get_serde();
     const auto field = arrow::field("payload", binary_variant_arrow_type(), true);
-    const Status status = paimon_arrow_write_converter().write_column(
+    const Status status = paimon::paimon_arrow_write_converter().write_column(
             type, *serde, *encoded, nullptr, field, arrow_builder.get(), 0, encoded->size(),
             cctz::utc_time_zone());
     EXPECT_EQ(status.code(), ErrorCode::NOT_IMPLEMENTED_ERROR);
